social_media_api/requirements.txt
acme==1.21.0
asgiref==3.8.1
blinker==1.4
certbot==1.21.0
certbot-nginx==1.21.0
certifi==2020.6.20
chardet==4.0.0
command-not-found==0.3
ConfigArgParse==1.5.3
configobj==5.0.6
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
dj-database-url==2.3.0
Django==5.1.4
django-filter==24.3
django-taggit==6.1.0
django_csp==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
josepy==1.10.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==24.2
parsedatetime==2.6
pillow==11.0.0
psycopg2-binary==2.9.10
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.2
systemd-python==234
typing_extensions==4.12.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
